requirements.txt
flask
flask-cors
pillow
opencv-python
numpy
piexif
